utils: split S3-over-RDMA rdma.cpp into focused units
Break the single rdma.{h,cpp} into three single-responsibility units, as
requested in review:
  - cuobj_client.{h,cpp}          cuObject buffer registration + token minting
  - s3_control_plane_http.{h,cpp} the signed, body-less control-plane GET/PUT
  - rdmaRetry.{h,cpp}             token-lifecycle + one-transient-retry wrappers

rdma_protocol.h (dependency-free wire helpers) is unchanged. rdma.h is now a
thin public umbrella that includes the four headers and carries the design
overview, so downstream consumers keep including "object/rdma/rdma.h".

Pure code relocation: no behavior change. Each new .cpp stays internally
guarded by HAVE_CUOBJ_CLIENT and compiles to nothing without the cuObjClient
library. rdmaRetry keeps a .cpp (rather than header-only) so the retry logic
stays in the compiled obj_utils library, as it is today.
